app/requirements.txt
fastapi
mlflow
pandas
boto3
python-dotenv
catboost
prometheus-fastapi-instrumentator
